refactor ClassSearch into one container component and one functional component
