+++ src/Pages/class-manage-user/ClassManage.js
@@ -1,0 +1,12 @@
+import React from "react";
+import ClassSearch from "../class-search/ClassSearch";
+
+const ClassManage = props => {
+	return (
+		<div>
+			<ClassSearch />
+		</div>
+	);
+};
+
+export default ClassManage;
