--- src/Pages/class-search/components/StyledClassSearch.js
@@ -1,14 +1,0 @@
-import React from "react";
-// import styled from "styled-components";
-
-export const searchIconSVG = (
-	<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<path fill-rule="evenodd" clip-rule="evenodd" d="M15.502 14H14.708L14.432 13.726C15.407 12.589 16 11.115 16 9.49997C16 5.90997 13.09 2.99997 9.5 2.99997C5.91 2.99997 3 5.90997 3 9.49997C3 13.09 5.91 16 9.5 16C11.115 16 12.588 15.408 13.725 14.434L14.001 14.708V15.5L18.999 20.491L20.49 19L15.502 14ZM9.5 14C7.014 14 5 11.986 5 9.49997C5 7.01497 7.014 4.99997 9.5 4.99997C11.985 4.99997 14 7.01497 14 9.49997C14 11.986 11.985 14 9.5 14Z" fill="black" fill-opacity="0.54" />
-	</svg>
-);
-
-export const locationIconSVG = (
-	<svg width="15" height="21" viewBox="0 0 15 21" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<path d="M7.5 0C3.35559 0 0 3.39223 0 7.57212C0 10.3768 0.937149 11.2533 5.90902 19.302C6.64281 20.4898 8.35801 20.4884 9.09094 19.302C14.0654 11.2491 15 10.3763 15 7.57212C15 3.38785 11.6401 0 7.5 0ZM4.375 7.57212C4.375 5.82962 5.7741 4.41707 7.5 4.41707C9.2259 4.41707 10.625 5.82962 10.625 7.57212C10.625 9.31461 9.2259 10.7272 7.5 10.7272C5.7741 10.7272 4.375 9.31461 4.375 7.57212Z" fill="#FF9233" />
-	</svg>
-);
